+++ pkg/controllers/sns.go
@@ -69,10 +69,10 @@ func SNS(c *fiber.Ctx) error {
 			if len(c.Queries()) > 0 {
 				overwriteVaule := c.Queries()
 				overwriteVaule["incident_source"] = "sns"
-				err = services.CreateIncident("", content, &overwriteVaule)
+				_, err = services.CreateIncident("", content, &overwriteVaule)
 			} else {
 				overwriteVaule := map[string]string{"incident_source": "sns"}
-				err = services.CreateIncident("", content, &overwriteVaule)
+				_, err = services.CreateIncident("", content, &overwriteVaule)
 			}
 
 			if err != nil {